[architecture] support rounds, use concurrency (#1)
This is an overhaul of the existing architecture.
Renderer is decoupled to refresh at a set interval rather than tied to game update.
The players follow the same approach.
Data structures to hold game data is slimmed down (and represented as arrays).

This will also eventually move in the direction of being a server-client program.